requirements.txt
langchain==0.2.10
# langchain-openai==0.1.17
python-dotenv==1.0.1
langchain-groq==0.1.6
langchain-community==0.2.10
langchain-chroma==0.1.2
tiktoken
chromadb
flask
GitPython
sentence-transformers==3.0.1
langchain-google-genai